+++ app/models/transfer/task_list.rb
@@ -6,6 +6,7 @@ class Transfer::TaskList < ::BaseTaskList
         tasks: [
           Transfer::Task::HandoverTaskForm,
           Transfer::Task::StakeholderKickOffTaskForm,
+          Transfer::Task::ConfirmHeadteacherContactTaskForm,
           Transfer::Task::MainContactTaskForm,
           Transfer::Task::RequestNewUrnAndRecordTaskForm,
           Transfer::Task::SponsoredSupportGrantTaskForm,